requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.4
python-dotenv>=0.21.0